Rename links store
